+++ android-gif-drawable/src/main/c/gif.h
@@ -138,6 +138,7 @@ typedef struct {
 	jlong position;
 	jbyte *bytes;
 	jlong capacity;
+	jobject bufferRef;
 } DirectByteBufferContainer;
 
 typedef struct {
@@ -224,4 +225,4 @@ uint_fast32_t seek(GifInfo *info, uint_fast32_t desiredIndex, void *pixels);
 
 void setGCBDefaults(GraphicsControlBlock *gcb);
 
-static GifInfo *createGifInfoFromFile(JNIEnv *env, FILE *file, const long sourceLength);
+static GifInfo *createGifInfoFromFile(JNIEnv *env, FILE *file, const long long sourceLength);