+++ android-gif-drawable/src/main/c/giflib/gifalloc.c
@@ -5,7 +5,7 @@
 ****************************************************************************/
 
 #include <stdlib.h>
-#include <stdio.h>
+#include <string.h>
 #include "gif_lib.h"
 
 //#define MAX(x, y)    (((x) > (y)) ? (x) : (y))